Rename everything to Qibocal
