--- src/qcvv/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-from .cli import command, live_plot, upload
-
-"""qcvv: Quantum Calibration Verification and Validation using Qibo."""
-import importlib.metadata as im
-
-__version__ = im.version(__package__)
