--- src/qcvv/calibrations/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from qcvv.calibrations.characterization.allXY import *
-from qcvv.calibrations.characterization.calibrate_qubit_states import *
-from qcvv.calibrations.characterization.flipping import *
-from qcvv.calibrations.characterization.qubit_spectroscopy import *
-from qcvv.calibrations.characterization.rabi_oscillations import *
-from qcvv.calibrations.characterization.ramsey import *
-from qcvv.calibrations.characterization.resonator_spectroscopy import *
-from qcvv.calibrations.characterization.t1 import *
-from qcvv.calibrations.protocols.test import *
