--- src/qcvv/plots/heatmaps.py
@@ -1,238 +1,0 @@
-# -*- coding: utf-8 -*-
-import os.path
-
-import plotly.graph_objects as go
-from plotly.subplots import make_subplots
-
-from qcvv.data import Dataset
-
-
-def frequency_flux_msr_phase(folder, routine, qubit, format):
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("frequency", "GHz"),
-            y=data.get_values("current", "A"),
-            z=data.get_values("MSR", "V"),
-            colorbar_x=0.45,
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("frequency", "GHz"),
-            y=data.get_values("current", "A"),
-            z=data.get_values("phase", "rad"),
-            colorbar_x=1.0,
-        ),
-        row=1,
-        col=2,
-    )
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Frequency (GHz)",
-        yaxis_title="Current (A)",
-        xaxis2_title="Frequency (GHz)",
-        yaxis2_title="Current (A)",
-    )
-    return fig
-
-
-def frequency_attenuation_msr_phase(folder, routine, qubit, format):
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("frequency", "GHz"),
-            y=data.get_values("attenuation", "dB"),
-            z=data.get_values("MSR", "V"),
-            colorbar_x=0.45,
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("frequency", "GHz"),
-            y=data.get_values("attenuation", "dB"),
-            z=data.get_values("phase", "rad"),
-            colorbar_x=1.0,
-        ),
-        row=1,
-        col=2,
-    )
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Frequency (GHz)",
-        yaxis_title="Attenuation (dB)",
-        xaxis2_title="Frequency (GHz)",
-        yaxis2_title="Attenuation (dB)",
-    )
-    return fig
-
-
-def frequency_flux_msr_phase__matrix(folder, routine, qubit, format):
-    fluxes = []
-    for i in range(25):  # FIXME: 25 is hardcoded
-        file = f"{folder}/data/{routine}/data_q{qubit}_f{i}.csv"
-        if os.path.exists(file):
-            fluxes += [i]
-
-    if len(fluxes) < 1:
-        nb = 1
-    else:
-        nb = len(fluxes)
-    fig = make_subplots(
-        rows=2,
-        cols=nb,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        x_title="Frequency (Hz)",
-        y_title="Current (A)",
-        shared_xaxes=True,
-        shared_yaxes=True,
-    )
-
-    for j in fluxes:
-        if j == fluxes[-1]:
-            showscale = True
-        else:
-            showscale = False
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}_f{j}")
-        fig.add_trace(
-            go.Heatmap(
-                x=data.get_values("frequency", "GHz"),
-                y=data.get_values("current", "A"),
-                z=data.get_values("MSR", "V"),
-                showscale=showscale,
-            ),
-            row=1,
-            col=j,
-        )
-        fig.add_trace(
-            go.Heatmap(
-                x=data.get_values("frequency", "GHz"),
-                y=data.get_values("current", "A"),
-                z=data.get_values("phase", "rad"),
-                showscale=showscale,
-            ),
-            row=2,
-            col=j,
-        )
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-    )
-    return fig
-
-
-def duration_gain_msr_phase(folder, routine, qubit, format):
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("duration", "ns"),
-            y=data.get_values("gain", "dimensionless"),
-            z=data.get_values("MSR", "V"),
-            colorbar_x=0.45,
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("duration", "ns"),
-            y=data.get_values("gain", "dimensionless"),
-            z=data.get_values("phase", "rad"),
-            colorbar_x=1.0,
-        ),
-        row=1,
-        col=2,
-    )
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="duration (ns)",
-        yaxis_title="gain (dimensionless)",
-        xaxis2_title="duration (ns)",
-        yaxis2_title="gain (dimensionless)",
-    )
-    return fig
-
-
-def duration_amplitude_msr_phase(folder, routine, qubit, format):
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("duration", "ns"),
-            y=data.get_values("amplitude", "dimensionless"),
-            z=data.get_values("MSR", "V"),
-            colorbar_x=0.45,
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Heatmap(
-            x=data.get_values("duration", "ns"),
-            y=data.get_values("amplitude", "dimensionless"),
-            z=data.get_values("phase", "rad"),
-            colorbar_x=1.0,
-        ),
-        row=1,
-        col=2,
-    )
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="duration (ns)",
-        yaxis_title="amplitude (dimensionless)",
-        xaxis2_title="duration (ns)",
-        yaxis2_title="amplitude (dimensionless)",
-    )
-    return fig
