+++ src/qibocal/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from .cli import command, live_plot, upload
+
+"""qibocal: Quantum Calibration Verification and Validation using Qibo."""
+import importlib.metadata as im
+
+__version__ = im.version(__package__)
