+++ src/qibocal/calibrations/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from qibocal.calibrations.characterization.allXY import *
+from qibocal.calibrations.characterization.calibrate_qubit_states import *
+from qibocal.calibrations.characterization.flipping import *
+from qibocal.calibrations.characterization.qubit_spectroscopy import *
+from qibocal.calibrations.characterization.rabi_oscillations import *
+from qibocal.calibrations.characterization.ramsey import *
+from qibocal.calibrations.characterization.resonator_spectroscopy import *
+from qibocal.calibrations.characterization.t1 import *
+from qibocal.calibrations.protocols.test import *
