+++ src/qibocal/cli/_base.py
@@ -1,0 +1,154 @@
+# -*- coding: utf-8 -*-
+"""Adds global CLI options."""
+import base64
+import pathlib
+import shutil
+import socket
+import subprocess
+import uuid
+from urllib.parse import urljoin
+
+import click
+from qibo.config import log, raise_error
+
+from qibocal.cli.builders import ActionBuilder
+
+CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
+
+# options for report upload
+UPLOAD_HOST = (
+    "qcvv@localhost"
+    if socket.gethostname() == "saadiyat"
+    else "qcvv@login.qrccluster.com"
+)
+TARGET_DIR = "qcvv-reports/"
+ROOT_URL = "http://login.qrccluster.com:9000/"
+
+
+@click.command(context_settings=CONTEXT_SETTINGS)
+@click.argument("runcard", metavar="RUNCARD", type=click.Path(exists=True))
+@click.option(
+    "folder",
+    "-o",
+    type=click.Path(),
+    help="Output folder. If not provided a standard name will generated.",
+)
+@click.option(
+    "force",
+    "-f",
+    is_flag=True,
+    help="Use --force option to overwrite the output folder.",
+)
+def command(runcard, folder, force=None):
+
+    """qibocal: Quantum Calibration Verification and Validation using Qibo.
+
+    Arguments:
+
+     - RUNCARD: runcard with declarative inputs.
+    """
+
+    builder = ActionBuilder(runcard, folder, force)
+    builder.execute()
+    builder.dump_report()
+
+
+@click.command(context_settings=CONTEXT_SETTINGS)
+@click.option(
+    "port",
+    "-p",
+    "--port",
+    default=8050,
+    type=int,
+    help="Localhost port to launch dash server.",
+)
+@click.option(
+    "debug",
+    "-d",
+    "--debug",
+    is_flag=True,
+    help="Launch server in debugging mode.",
+)
+def live_plot(port, debug):
+    """Real time plotting of calibration data on a dash server."""
+    import socket
+
+    from qibocal.web.app import app
+
+    # change port if it is already used
+    while True:
+        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+            if s.connect_ex(("localhost", port)) != 0:
+                break
+        port += 1
+
+    app.run_server(debug=debug, port=port)
+
+
+@click.command(context_settings=CONTEXT_SETTINGS)
+@click.argument("output_folder", metavar="FOLDER", type=click.Path(exists=True))
+def upload(output_folder):
+    """Uploads output folder to server"""
+
+    output_path = pathlib.Path(output_folder)
+
+    # check the rsync command exists.
+    if not shutil.which("rsync"):
+        raise_error(
+            RuntimeError,
+            "Could not find the rsync command. Please make sure it is installed.",
+        )
+
+    # check that we can authentica with a certificate
+    ssh_command_line = (
+        "ssh",
+        "-o",
+        "PreferredAuthentications=publickey",
+        "-q",
+        UPLOAD_HOST,
+        "exit",
+    )
+
+    str_line = " ".join(repr(ele) for ele in ssh_command_line)
+
+    log.info(f"Checking SSH connection to {UPLOAD_HOST}.")
+
+    try:
+        subprocess.run(ssh_command_line, check=True)
+    except subprocess.CalledProcessError as e:
+        raise RuntimeError(
+            (
+                "Could not validate the SSH key. "
+                "The command\n%s\nreturned a non zero exit status. "
+                "Please make sure that your public SSH key is on the server."
+            )
+            % str_line
+        ) from e
+    except OSError as e:
+        raise RuntimeError(
+            "Could not run the command\n{}\n: {}".format(str_line, e)
+        ) from e
+
+    log.info("Connection seems OK.")
+
+    # upload output
+    randname = base64.urlsafe_b64encode(uuid.uuid4().bytes).decode()
+    newdir = TARGET_DIR + randname
+
+    rsync_command = (
+        "rsync",
+        "-aLz",
+        "--chmod=ug=rwx,o=rx",
+        f"{output_path}/",
+        f"{UPLOAD_HOST}:{newdir}",
+    )
+
+    log.info(f"Uploading output ({output_path}) to {UPLOAD_HOST}")
+    try:
+        subprocess.run(rsync_command, check=True)
+    except subprocess.CalledProcessError as e:
+        msg = f"Failed to upload output: {e}"
+        raise RuntimeError(msg) from e
+
+    url = urljoin(ROOT_URL, randname)
+    log.info(f"Upload completed. The result is available at:\n{url}")
