+++ src/qibocal/fitting/utils.py
