+++ src/qibocal/plots/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from qibocal.plots.heatmaps import *
+from qibocal.plots.scatters import *
