+++ src/qibocal/web/static/styles.css
